change int to double during integration

+++ Assignment7/Assignment7/Assignment7/masterWorkerIntegration.cpp
@@ -58,7 +58,7 @@ int main(int argc, char * argv[]) {
 
 double function(double a, double b, int n, int start, int end) {
     //perform the integration
-    int sol = 0;
+    double sol = 0;
     for(int i = start; i <= end; i++) {
         
         double inside = (a + i * ( ( b - a ) / n ) ) ;
